requirements.txt
yfinance
pandas
numpy
matplotlib
scikit-learn
tensorflow-cpu
streamlit>=1.24.0
requests
plotly
protobuf==3.20.3